refactor: move annotator modes type to shared annotator
