+++ application/ui/src/features/annotator/annotator-canvas/annotator-canvas.tsx
@@ -7,9 +7,9 @@ import { ZoomTransform } from '../../../components/zoom/zoom-transform';
 import type { Media } from '../../../constants/shared-types';
 import { useAnnotationActions } from '../../../shared/annotator/annotation-actions-provider.component';
 import { useAnnotationVisibility } from '../../../shared/annotator/annotation-visibility-provider.component';
+import type { AnnotatorMode } from '../../../shared/annotator/annotator-mode';
 import { useSelectedAnnotations } from '../../../shared/annotator/select-annotation-provider.component';
 import { isVideo, isVideoFrame } from '../../../shared/media-item-utils';
-import { AnnotatorMode } from '../../dataset/media-preview/secondary-toolbar/annotator-modes/mode';
 import { Annotations } from '../annotations/annotations.component';
 import { VideoAnnotations } from '../annotations/video-annotations.component';
 import { useIsAnnotatorSceneBusy } from '../hooks/use-is-annotator-scene-busy';
