+++ application/ui/src/features/dataset/media-preview/annotator-providers.component.tsx
@@ -7,10 +7,10 @@ import { ZoomProvider } from '../../../components/zoom/zoom.provider';
 import type { AnnotationDTO, Media } from '../../../constants/shared-types';
 import { AnnotationActionsProvider } from '../../../shared/annotator/annotation-actions-provider.component';
 import { AnnotationVisibilityProvider } from '../../../shared/annotator/annotation-visibility-provider.component';
+import type { AnnotatorMode } from '../../../shared/annotator/annotator-mode';
 import { SelectAnnotationProvider } from '../../../shared/annotator/select-annotation-provider.component';
 import { AnnotatorLabelsProvider } from '../../annotator/annotator-labels-provider.component';
 import { CanvasSettingsProvider } from './primary-toolbar/settings/canvas-settings-provider.component';
-import type { AnnotatorMode } from './secondary-toolbar/annotator-modes/mode';
 
 type AnnotatorProvidersProps = {
     mediaItem: Media;
