+++ application/ui/src/shared/annotator/annotator-mode.ts
